# openapi/v1.yml
...
openapi: "3.0.0"
info:
  version: "1.0.0-alpha1"
  description: "API definition for Kleister, manage mod packs for Minecraft"
  title: "Kleister OpenAPI"
  contact:
    name: "Kleister"
    email: "kleister@webhippie.de"
    url: "https://kleister.eu"
  license:
    name: "Apache-2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
servers:
  - url: "https://try.kleister.eu/api/v1"
security:
  - Header: []
  - Bearer: []
  - Basic: []
  - Cookie: []
paths:
...
  /mods/{mod_id}/versions:
    get:
      summary: "Fetch all available versions for a mod"
      operationId: "ListVersions"
      tags:
        - "mod"
      parameters:
        - in: "path"
          name: "mod_id"
          description: "A mod identifier or slug"
          schema:
            type: "string"
          required: true
          x-example: "9hkaizdrwypr4ewjp4tu"
        - name: "search"
          in: "query"
          required: false
          schema:
            type: "string"
          description: "Search query"
          x-example: "John Doe"
        - name: "sort"
          in: "query"
          required: false
          schema:
            type: "string"
            default: "name"
            enum:
              - "name"
              - "public"
          description: "Sorting column"
          x-example: "name"
        - name: "order"
          in: "query"
          required: false
          schema:
            type: "string"
            default: "asc"
            enum:
              - "asc"
              - "desc"
          description: "Sorting order"
          x-example: "desc"
        - name: "limit"
          in: "query"
          required: false
          schema:
            type: "integer"
            default: 100
          description: "Paging limit"
          x-example: 100
        - name: "offset"
          in: "query"
          required: false
          schema:
            type: "integer"
            default: 0
          description: "Paging offset"
          x-example: 0
      responses:
        "200":
          description: "A collection of versions"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/versions"
        "403":
          description: "User is not authorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "404":
          description: "Mod not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "500":
          description: "Some internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        default:
          description: "Some error unrelated to the handler"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"

    post:
      summary: "Create a new version for a mod"
      operationId: "CreateVersion"
      tags:
        - "mod"
      parameters:
        - in: "path"
          name: "mod_id"
          description: "A mod identifier or slug"
          schema:
            type: "string"
          required: true
          x-example: "9hkaizdrwypr4ewjp4tu"
      requestBody:
        description: "The version data to create"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/version"
      responses:
        "200":
          description: "The created version data"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/version"
        "403":
          description: "User is not authorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "404":
          description: "Mod not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "422":
          description: "Failed to validate request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "500":
          description: "Some internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        default:
          description: "Some error unrelated to the handler"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
  /mods/{mod_id}/versions/{version_id}:
    get:
      summary: "Fetch a specific version for a mod"
      operationId: "ShowVersion"
      tags:
        - "mod"
      parameters:
        - in: "path"
          name: "mod_id"
          description: "A mod identifier or slug"
          schema:
            type: "string"
          required: true
          x-example: "9hkaizdrwypr4ewjp4tu"
        - in: "path"
          name: "version_id"
          description: "A version identifier or slug"
          schema:
            type: "string"
          required: true
          x-example: "rh5bogtvofnmoxglyjrm"
      responses:
        "200":
          description: "The fetched version details"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/version"
        "403":
          description: "User is not authorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "404":
          description: "Version or mod not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "500":
          description: "Some internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        default:
          description: "Some error unrelated to the handler"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"

    put:
      summary: "Update a specific version for a mod"
      operationId: "UpdateVersion"
      tags:
        - "mod"
      parameters:
        - in: "path"
          name: "mod_id"
          description: "A mod identifier or slug"
          schema:
            type: "string"
          required: true
          x-example: "9hkaizdrwypr4ewjp4tu"
        - in: "path"
          name: "version_id"
          description: "A version identifier or slug"
          schema:
            type: "string"
          required: true
          x-example: "rh5bogtvofnmoxglyjrm"
      requestBody:
        description: "The version data to update"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/version"
      responses:
        "200":
          description: "The updated version details"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/version"
        "403":
          description: "User is not authorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "404":
          description: "Version or mod not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "422":
          description: "Failed to validate request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "500":
          description: "Some internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        default:
          description: "Some error unrelated to the handler"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"

    delete:
      summary: "Delete a specific version for a mod"
      operationId: "DeleteVersion"
      tags:
        - "mod"
      parameters:
        - in: "path"
          name: "mod_id"
          description: "A mod identifier or slug"
          schema:
            type: "string"
          required: true
          x-example: "9hkaizdrwypr4ewjp4tu"
        - in: "path"
          name: "version_id"
          description: "A version identifier or slug"
          schema:
            type: "string"
          required: true
          x-example: "rh5bogtvofnmoxglyjrm"
      responses:
        "200":
          description: "Plain success message"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "403":
          description: "User is not authorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "404":
          description: "Version or mod not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "400":
          description: "Failed to delete the version"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "500":
          description: "Some internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        default:
          description: "Some error unrelated to the handler"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
  /mods/{mod_id}/versions/{version_id}/builds:
    get:
      summary: "Fetch all builds attached to version"
      operationId: "ListVersionBuilds"
      tags:
        - "mod"
      parameters:
        - in: "path"
          name: "mod_id"
          description: "A mod identifier or slug"
          schema:
            type: "string"
          required: true
          x-example: "9hkaizdrwypr4ewjp4tu"
        - in: "path"
          name: "version_id"
          description: "A version identifier or slug"
          schema:
            type: "string"
          required: true
          x-example: "rh5bogtvofnmoxglyjrm"
        - name: "search"
          in: "query"
          required: false
          schema:
            type: "string"
          description: "Search query"
          x-example: "John Doe"
        - name: "sort"
          in: "query"
          required: false
          schema:
            type: "string"
            default: "name"
            enum:
              - "name"
              - "public"
          description: "Sorting column"
          x-example: "name"
        - name: "order"
          in: "query"
          required: false
          schema:
            type: "string"
            default: "asc"
            enum:
              - "asc"
              - "desc"
          description: "Sorting order"
          x-example: "desc"
        - name: "limit"
          in: "query"
          required: false
          schema:
            type: "integer"
            default: 100
          description: "Paging limit"
          x-example: 100
        - name: "offset"
          in: "query"
          required: false
          schema:
            type: "integer"
            default: 0
          description: "Paging offset"
          x-example: 0
      responses:
        "200":
          description: "A collection of version builds"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/version_builds"
        "403":
          description: "User is not authorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "404":
          description: "Version or mod not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "500":
          description: "Some internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        default:
          description: "Some error unrelated to the handler"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"

    post:
      summary: "Attach a build to a version"
      operationId: "AttachVersionToBuild"
      tags:
        - "mod"
      parameters:
        - in: "path"
          name: "mod_id"
          description: "A mod identifier or slug"
          schema:
            type: "string"
          required: true
          x-example: "9hkaizdrwypr4ewjp4tu"
        - in: "path"
          name: "version_id"
          description: "A version identifier or slug"
          schema:
            type: "string"
          required: true
          x-example: "rh5bogtvofnmoxglyjrm"
      requestBody:
        description: "The version build data to attach"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/version_build_params"
      responses:
        "200":
          description: "Plain success message"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "403":
          description: "User is not authorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "404":
          description: "Build, version or mod not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "412":
          description: "Build is already attached"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "422":
          description: "Failed to validate request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "500":
          description: "Some internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        default:
          description: "Some error unrelated to the handler"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"

    delete:
      summary: "Unlink a build from a version"
      operationId: "DeleteVersionFromBuild"
      tags:
        - "mod"
      parameters:
        - in: "path"
          name: "mod_id"
          description: "A mod identifier or slug"
          schema:
            type: "string"
          required: true
          x-example: "9hkaizdrwypr4ewjp4tu"
        - in: "path"
          name: "version_id"
          description: "A version identifier or slug"
          schema:
            type: "string"
          required: true
          x-example: "rh5bogtvofnmoxglyjrm"
      requestBody:
        description: "The version build data to unlink"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/version_build_params"
      responses:
        "200":
          description: "Plain success message"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "403":
          description: "User is not authorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "404":
          description: "Build, version or mod not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "412":
          description: "Build is not attached"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "500":
          description: "Some internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        default:
          description: "Some error unrelated to the handler"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
...
components:
  securitySchemes:
    Header:
      type: apiKey
      in: header
      name: X-API-Key
    Bearer:
      type: http
      scheme: bearer
    Basic:
      type: http
      scheme: basic
    Cookie:
      type: apiKey
      in: header
      name: Cookie
  schemas:
...
    mod:
      title: "Mod"
      description: "Model to represent mod"
      type: "object"
      properties:
        id:
          type: "string"
          readOnly: true
        slug:
          type: "string"
          x-omitempty: true
          x-nullable: true
        name:
          type: "string"
          x-omitempty: true
          x-nullable: true
        side:
          type: "string"
          enum:
            - "both"
            - "server"
            - "client"
          x-omitempty: true
          x-nullable: true
        description:
          type: "string"
          x-omitempty: true
          x-nullable: true
        author:
          type: "string"
          x-omitempty: true
          x-nullable: true
        website:
          type: "string"
          x-omitempty: true
          x-nullable: true
        donate:
          type: "string"
          x-omitempty: true
          x-nullable: true
        public:
          type: "boolean"
          x-omitempty: true
          x-nullable: true
        created_at:
          type: "string"
          format: "date-time"
          readOnly: true
        updated_at:
          type: "string"
          format: "date-time"
          readOnly: true
    versions:
      title: "Versions"
      description: "Model to represent list of versions"
      type: "object"
      properties:
        total:
          type: integer
          format: int64
        mod:
          readOnly: true
          $ref: "#/components/schemas/mod"
        versions:
          type: "array"
          items:
            $ref: "#/components/schemas/version"
    version:
      title: "Version"
      description: "Model to represent version"
      type: "object"
      properties:
        id:
          type: "string"
          readOnly: true
        file:
          x-omitempty: true
          x-nullable: true
          readOnly: true
          $ref: "#/components/schemas/version_file"
        mod:
          x-omitempty: true
          x-nullable: true
          readOnly: true
          $ref: "#/components/schemas/mod"
        slug:
          type: "string"
          x-omitempty: true
          x-nullable: true
        name:
          type: "string"
          x-omitempty: true
          x-nullable: true
        public:
          type: "boolean"
          x-omitempty: true
          x-nullable: true
        created_at:
          type: "string"
          format: "date-time"
          readOnly: true
        updated_at:
          type: "string"
          format: "date-time"
          readOnly: true
    version_file:
      title: "Version File"
      description: "Model to represent version file"
      type: "object"
      properties:
        id:
          type: "string"
          readOnly: true
        slug:
          type: "string"
          x-omitempty: true
          x-nullable: true
        content_type:
          type: "string"
          x-omitempty: true
          x-nullable: true
        md5:
          type: "string"
          x-omitempty: true
          x-nullable: true
        path:
          type: "string"
          x-omitempty: true
          x-nullable: true
        url:
          type: "string"
          x-omitempty: true
          x-nullable: true
        upload:
          type: "string"
          x-omitempty: true
          x-nullable: true
        created_at:
          type: "string"
          format: "date-time"
          readOnly: true
        updated_at:
          type: "string"
          format: "date-time"
          readOnly: true
...
    version_builds:
      title: "Version Builds"
      description: "Model to represent version builds"
      type: "object"
      properties:
        mod:
          readOnly: true
          $ref: "#/components/schemas/mod"
        version:
          readOnly: true
          $ref: "#/components/schemas/version"
        total:
          type: integer
          format: int64
        builds:
          type: "array"
          items:
            $ref: "#/components/schemas/build_version"
    version_build_params:
      title: "Version Build Params"
      description: "Parameters to attach or unlink version build"
      type: "object"
      required:
        - "pack"
        - "build"
      properties:
        pack:
          type: "string"
        build:
          type: "string"
    build_version:
      title: "Build Version"
      description: "Model to represent build version"
      type: "object"
      required:
        - "build_id"
        - "version_id"
      properties:
        build_id:
          type: "string"
        build:
          x-omitempty: true
          x-nullable: true
          readOnly: true
          $ref: "#/components/schemas/build"
        version_id:
          type: "string"
        version:
          x-omitempty: true
          x-nullable: true
          readOnly: true
          $ref: "#/components/schemas/version"
        created_at:
          type: "string"
          format: "date-time"
          readOnly: true
        updated_at:
          type: "string"
          format: "date-time"
          readOnly: true
    notification:
      title: "Notification"
      description: "Generic response for errors and validations"
      type: "object"
      properties:
        status:
          type: "integer"
          x-omitempty: true
        message:
          type: "string"
          x-omitempty: true
        errors:
          type: "array"
          items:
            $ref: "#/components/schemas/validation"
    validation:
      title: "Validation"
      description: "General structure to show validation errors"
      type: "object"
      properties:
        field:
          type: "string"
        message:
          type: "string"
...